+++ packages/kiarina-agi-data-builder/src/kiarina/agi/file_info_builder/_types/file_info_spec.py
@@ -36,3 +36,4 @@ class FileInfoSpec(TypedDict):
     # pdf
     start_page: NotRequired[int]
     end_page: NotRequired[int]
+    analysis_dpi: NotRequired[int]