+++ packages/kiarina-agi-data-builder/tests/file_info_builder_impl/pdf/_models/test_pdf_file_info_builder.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+from kiarina.agi.file_info import PDFFileInfo
 from kiarina.agi.file_info_builder import build_file_info
 from kiarina.agi.run_context import RunContext
 from kiarina.utils.file.asyncio import read_file
@@ -17,3 +18,5 @@ async def test_pdf_file_builder(run_context: RunContext, pdf_file_path: Path) ->
 
     print("PDFFileInfo:")
     print(file.file_info.model_dump_json(indent=2))
+    assert isinstance(file.file_info, PDFFileInfo)
+    assert file.file_info.analysis_dpi == 144